[SERVICE-539] Move `WindowIdentity`
