+++ src/provider/snapanddock/SnapService.ts
@@ -4,7 +4,7 @@ import {DesktopModel} from '../model/DesktopModel';
 import {DesktopSnapGroup} from '../model/DesktopSnapGroup';
 import {DesktopWindow} from '../model/DesktopWindow';
 import {Target} from '../WindowHandler';
-import {WindowIdentity} from '../utils/identity';
+import {WindowIdentity} from '../model/Identity';
 
 import {EXPLODE_MOVE_SCALE, UNDOCK_MOVE_DISTANCE} from './Constants';
 import {Resolver, SnapTarget} from './Resolver';
